% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{create_pathway_sets}
\alias{create_pathway_sets}
\title{Create Pathway Sets}
\usage{
create_pathway_sets(merged_data, classification_column)
}
\arguments{
\item{merged_data}{Data frame with lipids and classifications}

\item{classification_column}{Column name for classification}
}
\value{
Named list of pathway sets
}
\description{
Create Pathway Sets
}
