% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{fix_sample_alignment}
\alias{fix_sample_alignment}
\title{Align Samples Between Data Matrix and Metadata}
\usage{
fix_sample_alignment(data_matrix, metadata)
}
\arguments{
\item{data_matrix}{Numeric matrix (features as rows, samples as columns).}

\item{metadata}{Metadata data frame.}
}
\value{
Named list with aligned \code{data_matrix} and \code{metadata}.
}
\description{
Align Samples Between Data Matrix and Metadata
}
