% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{perform_enrichment_analysis}
\alias{perform_enrichment_analysis}
\title{Perform Enrichment Analysis}
\usage{
perform_enrichment_analysis(
  results_list,
  classification_data,
  min_set_size = 5,
  max_set_size = 500,
  custom_sets = NULL
)
}
\arguments{
\item{results_list}{List of differential analysis results}

\item{classification_data}{Lipid classification data frame}

\item{min_set_size}{Minimum pathway set size}

\item{max_set_size}{Maximum pathway set size}

\item{custom_sets}{Optional named list of custom lipid sets}
}
\value{
List containing GSEA results
}
\description{
Perform Enrichment Analysis
}
\examples{
d <- load_lipidomics_data_from_df(generate_example_data())
norm <- apply_normalizations(d$numeric_data, c("TIC", "Log2"))
cls <- classify_lipids(colnames(norm))
res <- perform_differential_analysis(norm, d$metadata, "Sample Group",
  contrasts_list = NULL, method = "limma"
)
enrich <- perform_enrichment_analysis(res$results, cls, min_set_size = 3)
names(enrich)
}
