% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{register_select_all_observers}
\alias{register_select_all_observers}
\title{Register Select-All and Deselect-All Observers for a CheckboxGroupInput}
\usage{
register_select_all_observers(session, input, inputId, choices_reactive)
}
\arguments{
\item{session}{Shiny session object.}

\item{input}{Shiny input object.}

\item{inputId}{The checkbox group input ID (buttons follow naming convention
\code{<inputId>_select_all} and \code{<inputId>_deselect_all}).}

\item{choices_reactive}{A reactive expression (or function) that returns the
current set of choices.}
}
\description{
Register Select-All and Deselect-All Observers for a CheckboxGroupInput
}
\keyword{internal}
