% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_normalizations}
\alias{apply_normalizations}
\title{Apply Multiple Normalizations}
\usage{
apply_normalizations(data, methods)
}
\arguments{
\item{data}{Numeric data matrix}

\item{methods}{Vector of normalization methods to apply in order}
}
\value{
Normalized data matrix
}
\description{
Apply Multiple Normalizations
}
