% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify_lipids}
\alias{classify_lipids}
\title{Classify Lipids Based on Names}
\usage{
classify_lipids(lipid_names)
}
\arguments{
\item{lipid_names}{Vector of lipid names}
}
\value{
Data frame with classification results
}
\description{
Classify Lipids Based on Names
}
