% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_enrichment_barplot}
\alias{create_enrichment_barplot}
\title{Create an Enrichment Barplot}
\usage{
create_enrichment_barplot(
  enrichment_data,
  title = "Enrichment Analysis",
  max_pathways = 15
)
}
\arguments{
\item{enrichment_data}{Data frame of fgsea results.}

\item{title}{Plot title.}

\item{max_pathways}{Maximum number of pathways (top by p-value).}
}
\value{
A \code{ggplot2} object.
}
\description{
Create an Enrichment Barplot
}
\examples{
d <- load_lipidomics_data_from_df(generate_example_data())
norm <- apply_normalizations(d$numeric_data, c("TIC", "Log2"))
cls <- classify_lipids(colnames(norm))
res <- perform_differential_analysis(norm, d$metadata, "Sample Group",
  contrasts_list = NULL, method = "limma"
)
enrich <- perform_enrichment_analysis(res$results, cls, min_set_size = 3)
p <- create_enrichment_barplot(enrich[[1]][["LipidGroup"]])
print(p)
}
