% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{create_enrichment_barplot}
\alias{create_enrichment_barplot}
\title{Create Enrichment Barplot - FIXED terminology}
\usage{
create_enrichment_barplot(
  enrichment_data,
  title = "Enrichment Analysis",
  max_pathways = 15
)
}
\arguments{
\item{enrichment_data}{Enrichment results data frame}

\item{title}{Plot title}

\item{max_pathways}{Maximum number of pathways to show}
}
\value{
ggplot object
}
\description{
Create Enrichment Barplot - FIXED terminology
}
