% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determine_saturation}
\alias{determine_saturation}
\title{Enhanced Saturation Determination Function}
\usage{
determine_saturation(lipid_name)
}
\arguments{
\item{lipid_name}{Single lipid name}
}
\value{
Saturation classification (SFA, MUFA, PUFA, or Unclassified)
}
\description{
Enhanced Saturation Determination Function
}
