% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{fix_sample_alignment}
\alias{fix_sample_alignment}
\title{Fix Sample Alignment between data and metadata}
\usage{
fix_sample_alignment(data_matrix, metadata)
}
\arguments{
\item{data_matrix}{Data matrix (features as rows, samples as columns)}

\item{metadata}{Metadata data frame}
}
\value{
List with aligned data_matrix and metadata
}
\description{
Fix Sample Alignment between data and metadata
}
