% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_lipid_classification}
\alias{get_lipid_classification}
\title{Get Lipid Classification from Names}
\usage{
get_lipid_classification(lipid_names)
}
\arguments{
\item{lipid_names}{Vector of lipid names to classify}
}
\value{
Data frame with lipid classifications (Lipid, LipidGroup, LipidType, Saturation)
}
\description{
Wrapper function to classify lipids and return the classification data frame.
This function is exported for use in scripts.
}
\examples{
lipids <- c("PC 16:0_18:1", "PE 18:0_20:4", "TG 16:0_18:1_20:4")
classification <- get_lipid_classification(lipids)
print(classification)
}
