% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_normalization_methods}
\alias{get_normalization_methods}
\title{Return Available Normalization Method Names}
\usage{
get_normalization_methods()
}
\value{
Character vector of normalization method names supported by
  \code{apply_normalizations}.
}
\description{
Return Available Normalization Method Names
}
\examples{
get_normalization_methods()
}
