% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LIPIDIFy-package.R
\docType{package}
\name{LIPIDIFy-package}
\alias{LIPIDIFy}
\alias{LIPIDIFy-package}
\title{LIPIDIFy: Comprehensive Lipidomics Data Analysis with Interactive Visualization}
\description{
LIPIDIFy provides a comprehensive toolkit for end-to-end lipidomics data
analysis, including missing value imputation, batch effect correction,
normalization, differential abundance analysis using limma and edgeR,
gene set enrichment analysis, and extensive visualization capabilities.
Lipid names are automatically classified by class, subclass, and
fatty-acid saturation. The package offers both an interactive Shiny
interface for bench biologists (see \code{\link{launch_lipidomics_app}})
and fully scriptable R functions for bioinformaticians. It supports
flexible custom lipid classification schemes and user-defined enrichment
sets.
}
\section{Getting started}{

See \code{vignette("LIPIDIFy")} for a full walkthrough of the
programmatic workflow.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fayrouzhammal/LIPIDIFy}
  \item Report bugs at \url{https://github.com/fayrouzhammal/LIPIDIFy/issues}
}

}
\author{
\strong{Maintainer}: Fayrouz Hammal \email{fayrouz.hammal@petermac.org} (\href{https://orcid.org/0000-0002-7612-4953}{ORCID})

Authors:
\itemize{
  \item Fayrouz Hammal \email{fayrouz.hammal@petermac.org} (\href{https://orcid.org/0000-0002-7612-4953}{ORCID})
}

}
\keyword{internal}
