% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_lipidomics_data}
\alias{load_lipidomics_data}
\title{Load Lipidomics Data}
\usage{
load_lipidomics_data(
  file_path,
  metadata_columns = c("Sample Name", "Sample Group", "Tumour ID", "Weight (mg)")
)
}
\arguments{
\item{file_path}{Path to the lipidomics data file}

\item{metadata_columns}{Vector of metadata column names}
}
\value{
List containing data matrix and metadata
}
\description{
Load Lipidomics Data
}
