% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_lipidomics_data}
\alias{normalize_lipidomics_data}
\title{Normalize Lipidomics Data}
\usage{
normalize_lipidomics_data(data, methods = c("TIC", "Log2"))
}
\arguments{
\item{data}{Numeric data matrix}

\item{methods}{Vector of normalization methods}
}
\value{
Normalized data matrix
}
\description{
Normalize Lipidomics Data
}
