% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_mean}
\alias{normalize_mean}
\title{Mean Normalization}
\usage{
normalize_mean(data)
}
\arguments{
\item{data}{Numeric data matrix (samples as rows)}
}
\value{
Mean normalized data
}
\description{
Mean Normalization
}
