% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_median}
\alias{normalize_median}
\title{Median Normalization}
\usage{
normalize_median(data)
}
\arguments{
\item{data}{Numeric data matrix (samples as rows)}
}
\value{
Median normalized data
}
\description{
Median Normalization
}
