% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_pqn}
\alias{normalize_pqn}
\title{PQN Normalization}
\usage{
normalize_pqn(data)
}
\arguments{
\item{data}{Numeric data matrix (samples as rows)}
}
\value{
PQN normalized data
}
\description{
PQN Normalization
}
