% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_quantile}
\alias{normalize_quantile}
\title{Quantile Normalization}
\usage{
normalize_quantile(data)
}
\arguments{
\item{data}{Numeric data matrix}
}
\value{
Quantile normalized data
}
\description{
Quantile Normalization
}
