% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_tic}
\alias{normalize_tic}
\title{TIC Normalization}
\usage{
normalize_tic(data)
}
\arguments{
\item{data}{Numeric data matrix (samples as rows)}
}
\value{
TIC normalized data
}
\description{
TIC Normalization
}
