% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize_vsn}
\alias{normalize_vsn}
\title{VSN Normalization (simplified)}
\usage{
normalize_vsn(data)
}
\arguments{
\item{data}{Numeric data matrix}
}
\value{
VSN normalized data
}
\description{
VSN Normalization (simplified)
}
