% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{perform_plsda}
\alias{perform_plsda}
\title{Perform PLS-DA Analysis (FIXED VERSION)}
\usage{
perform_plsda(data_matrix, metadata, group_column = "Sample Group", n_comp = 2)
}
\arguments{
\item{data_matrix}{Data matrix (samples as rows)}

\item{metadata}{Metadata data frame}

\item{group_column}{Group column name}

\item{n_comp}{Number of components}
}
\value{
List containing PLS-DA results and plot
}
\description{
Perform PLS-DA Analysis (FIXED VERSION)
}
\examples{
d <- load_lipidomics_data_from_df(generate_example_data())
norm <- apply_normalizations(d$numeric_data, c("TIC", "Log2"))
res <- perform_plsda(norm, d$metadata, "Sample Group")
names(res)
}
