% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_saturation_classification}
\alias{test_saturation_classification}
\title{Test function to check saturation classification}
\usage{
test_saturation_classification(test_lipids = NULL)
}
\arguments{
\item{test_lipids}{Vector of test lipid names}
}
\value{
Data frame showing classification results
}
\description{
Test function to check saturation classification
}
