% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{visualize_raw_data}
\alias{visualize_raw_data}
\title{Visualize Raw Data}
\usage{
visualize_raw_data(data_list, plot_type = "boxplot")
}
\arguments{
\item{data_list}{List containing data components from load_lipidomics_data}

\item{plot_type}{Type of plot ("boxplot", "density", "histogram")}
}
\value{
ggplot object
}
\description{
Visualize Raw Data
}
